+++ sql/product/product.sql
@@ -2,6 +2,7 @@
 DROP TABLE product CASCADE CONSTRAINT;
 DROP TABLE product_image CASCADE CONSTRAINT;
 DROP TABLE product_course_point CASCADE CONSTRAINT;
+
 -- 기존 시퀀스 삭제
 DROP SEQUENCE PRODUCT_PRODUCT_ID_SEQ;
 DROP SEQUENCE IMAGE_IMAGE_ID_SEQ;
