--- src/main/resources/static/css/product/product_area.css
@@ -1,123 +1,0 @@
-body {
-  background: #fff;
-  margin: 0;
-  font-family: 'Pretendard', 'Noto Sans KR', sans-serif;
-}
-a {
-    text-decoration: none;
-    color: inherit;
-}
-.area-container {
-  width: 1100px;
-  margin: 0 auto;
-  padding-top: 40px;
-}
-.area-breadcrumb {
-  color: #888;
-  font-size: 14px;
-  margin-bottom: 16px;
-}
-.area-breadcrumb-current {
-  color: #0099ad;
-  font-weight: 500;
-}
-.area-title {
-  font-size: 32px;
-  font-weight: 700;
-  margin-bottom: 32px;
-}
-.area-keywords-section {
-  background: #fafbfc;
-  border-radius: 12px;
-  padding: 28px 32px 24px 32px;
-  margin-bottom: 32px;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.03);
-}
-.area-keywords-title {
-  font-size: 20px;
-  font-weight: bold;
-  margin-bottom: 18px;
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-.hot-badge {
-  background: #00bfae;
-  color: #fff;
-  font-size: 12px;
-  font-weight: bold;
-  border-radius: 8px;
-  padding: 2px 8px;
-  margin-left: 6px;
-}
-.area-keywords-list {
-  display: flex;
-  gap: 18px;
-}
-.area-keyword {
-  border: 2px solid #00bfae;
-  border-radius: 12px;
-  padding: 12px 28px;
-  font-size: 20px;
-  font-weight: 500;
-  background: #fff;
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-.emoji {
-  font-size: 26px;
-}
-.area-main-section {
-  display: flex;
-  margin-top: 32px;
-}
-.area-category-list {
-  min-width: 120px;
-  margin-right: 32px;
-  display: flex;
-  flex-direction: column;
-  gap: 18px;
-}
-.area-category {
-  font-size: 18px;
-  font-weight: 600;
-  color: #222;
-  margin-bottom: 2px;
-}
-.area-content {
-  flex: 1;
-}
-.area-filter-box {
-  background: #fcfcfc;
-  border: 1.5px solid #e0e0e0;
-  border-radius: 12px;
-  padding: 24px 24px 18px 24px;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.03);
-}
-.area-filter-row {
-  display: flex;
-  gap: 12px;
-  margin-bottom: 12px;
-}
-.area-filter-box select {
-  padding: 8px 16px;
-  border: 1px solid #e0e0e0;
-  border-radius: 8px;
-  font-size: 15px;
-  background: #fff;
-}
-.area-keyword-input {
-  width: 100%;
-  margin: 10px 0 0 0;
-  padding: 10px 14px;
-  border: 1px solid #e0e0e0;
-  border-radius: 8px;
-  font-size: 15px;
-  background: #fff;
-}
-.area-filter-desc {
-  color: #888;
-  font-size: 14px;
-  margin-top: 10px;
-}
