--- src/main/resources/static/js/product/product_area.js
@@ -1,3 +1,0 @@
-
-
-
