+++ src/main/resources/templates/fragments/header.html
@@ -51,7 +51,7 @@
     <nav class="header-nav">
       <ul>
         <li><a th:href="@{/menu}">전체메뉴</a></li>
-        <li><a th:href="@{/theme}">테마여행</a></li>
+        <li><a th:href="@{/product/list}">테마여행</a></li>
         <li><a th:href="@{/bbs}">커뮤니티</a></li>
         <li><a th:href="@{/air}">항공</a></li>
         <li><a th:href="@{/hotel}">호텔</a></li>
