+++ src/main/resources/templates/mypage/sellerpage/editSellerPage.html
@@ -37,7 +37,7 @@
             <ul>
               <li><a href="#">상품</a>
                 <ul>
-                  <li><a th:href="@{/product/list}">내 상품 목록</a></li>
+                  <li><a th:href="@{/product/manage}">내 상품 목록</a></li>
                   <li><a th:href="@{/product/upload}">상품 등록</a></li>
                 </ul>
               </li>
