+++ src/main/resources/templates/mypage/sellerpage/viewSellerPage.html
@@ -38,8 +38,8 @@
               <ul>
                 <li><strong>상품</strong>
                   <ul>
-                    <li><a th:href="@{/seller/products}">내 상품 목록</a></li>
-                    <li><a th:href="@{/seller/products/new}">상품 등록</a></li>
+                    <li><a th:href="@{/product/manage}">내 상품 목록</a></li>
+                    <li><a th:href="@{/product/upload}">상품 등록</a></li>
                   </ul>
                 </li>
                 <li><strong>판매 현황</strong>
