+++ src/main/resources/templates/product/product_enroll.html
@@ -7,7 +7,6 @@
   <title>MO:SI 판매자 마이페이지</title>
   <!-- <link rel="stylesheet" href="/src/main/resources/static/css/layout.css">
   <link rel="stylesheet" href="/src/main/resources/static/css/sellerpage/sellerMypageHome.css"> -->
-  <link rel="stylesheet" th:href="@{/css/common.css}">
   <link rel="stylesheet" th:href="@{/css/layout.css}">
   <link rel="stylesheet" th:href="@{/css/common.css}">
   <link rel="stylesheet" th:href="@{/css/sellerpage/sellerMypageHome.css}">
@@ -43,7 +42,7 @@
               <ul>
                 <li><a href="#">상품</a>
                   <ul>
-                    <li><a th:href="@{/product/list}">내 상품 목록</a></li>
+                    <li><a th:href="@{/product/manage}">내 상품 목록</a></li>
                     <li><a href="#">상품 등록</a></li>
                   </ul>
                 </li>
@@ -272,7 +271,7 @@
     </div>
     <footer th:replace="~{fragments/footer :: footer}"></footer>
   </div>
-  <script th:src="@{/js/product/enroll.js}"></script>
+  <script th:src="@{/js/product/product_enroll.js}"></script>
   <script th:src="@{/js/sellerpage/sellerMypageHome.js}"></script>
 </body>
