+++ src/main/resources/templates/product/product_managing.html
@@ -42,7 +42,7 @@
               <ul>
                 <li><a href="#">상품</a>
                   <ul>
-                    <li><a th:href="@{/product/list}">내 상품 목록</a></li>
+                    <li><a th:href="@{/product/manage}">내 상품 목록</a></li>
                     <li><a th:href="@{/product/upload}">상품 등록</a></li>
                   </ul>
                 </li>
@@ -96,7 +96,7 @@
               </div>
               <div class="overall-info">
                 <div th:text="|총 ${totalCount}개 상품|"></div>
-                <form th:action="@{/product/list}" method="get" id="statusForm">
+                <form th:action="@{/product/manage}" method="get" id="statusForm">
                   <input type="hidden" name="page" th:value="${currentPage}" />
                   <input type="hidden" name="size" th:value="5" />  <!-- size는 고정하거나 동적으로 설정 -->
 
@@ -162,12 +162,12 @@
               <!-- 페이지네이션 -->
               <div class="pagination">
                 <a th:if="${currentPage > 1}"
-                   th:href="@{/product/list(page=${currentPage - 1}, size=${5}, status=${selectedStatus})}">이전</a>
+                   th:href="@{/product/manage(page=${currentPage - 1}, size=${5}, status=${selectedStatus})}">이전</a>
 
                 <span th:text="${currentPage} + ' / ' + ${totalPages}"></span>
 
                 <a th:if="${currentPage < totalPages}"
-                   th:href="@{/product/list(page=${currentPage + 1}, size=${5}, status=${selectedStatus})}">다음</a>
+                   th:href="@{/product/manage(page=${currentPage + 1}, size=${5}, status=${selectedStatus})}">다음</a>
               </div>
             </div>
           </div>
